+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Convert Call with missing argument list and semantic information
 
 ### C# -> VB
-
+* Convert "is" expression ([#427](https://github.com/icsharpcode/CodeConverter/pull/427))
 
 ## [7.3.0] - 2019-11-25